Move tests to py_kwiver
