+++ py_kwiver/vital/tests/helpers.py
