+++ py_kwiver/vital/tests/test_camera_map.py
