+++ py_kwiver/vital/tests/test_color.py
