+++ py_kwiver/vital/tests/test_config.py
