+++ py_kwiver/vital/tests/test_covariance.py
