+++ py_kwiver/vital/tests/test_descriptor.py
