+++ py_kwiver/vital/tests/test_descriptor_set.py
