+++ py_kwiver/vital/tests/test_eigen_numpy.py
