+++ py_kwiver/vital/tests/test_feature.py
