+++ py_kwiver/vital/tests/test_homography.py
