+++ py_kwiver/vital/tests/test_image.py
