+++ py_kwiver/vital/tests/test_image_container.py
