+++ py_kwiver/vital/tests/test_rotation.py
