+++ py_kwiver/vital/tests/test_similarity.py
