+++ py_kwiver/vital/tests/test_track.py
