+++ py_kwiver/vital/tests/test_track_state.py
